@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import token from "../utils/token";
+import token from "../lib/token";
 
 const Logout = ({ setUser }) => {
   const handleLogout = (event) => {
@@ -12,7 +12,7 @@ const Logout = ({ setUser }) => {
   };
 
   return (
-    <form onSubmit={handleLogout}>
+    <form  className="logout" onSubmit={handleLogout}>
       <span className="logout">
         <button type="submit">Logout</button>
       </span>
