@@ -1,14 +1,60 @@
 import React from "react";
 import PropTypes from "prop-types";
+import { useTranslation } from "react-i18next";
 
 import Toggleable from "./Toggleable";
 import Login from "./Login";
 import Logout from "./Logout";
 import User from "./User";
 
-const TopMenu = ({ user, setUser }) => {
+const TopMenu = ({ user, setUser, lang }) => {
+  const { i18n } = useTranslation();
+
+  const handleLangChange = (event) => {
+    const lang = event.target.value;
+    i18n.changeLanguage(lang);
+    window.localStorage.setItem("lang", lang);
+  };
+
   const renderUserInfo = () => {
-    return <User user={user} />;
+    return (
+      <>
+        <User user={user} />{" "}
+        <span className="lang">
+          <form>
+            <label className={lang === "en" ? "checked" : ""}>
+              <input
+                type="radio"
+                value="en"
+                checked={lang === "en"}
+                onChange={handleLangChange}
+              />
+              English
+            </label>
+            |
+            <label className={lang === "fi" ? "checked" : ""}>
+              <input
+                type="radio"
+                value="fi"
+                checked={lang === "fi"}
+                onChange={handleLangChange}
+              />
+              Suomi
+            </label>
+            |
+            <label className={lang === "ja" ? "checked" : ""}>
+              <input
+                type="radio"
+                value="ja"
+                checked={lang === "ja"}
+                onChange={handleLangChange}
+              />
+              日本語
+            </label>
+          </form>
+        </span>
+      </>
+    );
   };
   const renderContent = () => {
     if (user) {
@@ -36,5 +82,6 @@ const TopMenu = ({ user, setUser }) => {
 TopMenu.propTypes = {
   user: PropTypes.object,
   setUser: PropTypes.func,
+  lang: PropTypes.string.isRequired,
 };
 export default TopMenu;
