@@ -3,17 +3,16 @@ const cors = require("cors");
 const compression = require("compression");
 require("express-async-errors");
 
-const config = require("./utils/config");
+const config = require("./lib/config");
 
 const tokens = require("./controllers/tokens");
 const users = require("./controllers/users");
-const stats = require("./controllers/stats");
 const galleries = require("./controllers/galleries");
 const photos = require("./controllers/photos");
 const galleryPhotos = require("./controllers/gallery-photos");
 
-const middleware = require("./utils/middleware");
-const logger = require("./utils/logger");
+const middleware = require("./lib/middleware");
+const logger = require("./lib/logger");
 
 const app = express();
 app.use(cors());
@@ -32,7 +31,6 @@ const registerPreProcessors = () => {
 const registerRoutes = () => {
   app.use("/api/tokens", tokens.router);
   app.use("/api/users", users.router);
-  app.use("/api/stats", stats.router);
   app.use("/api/galleries", galleries.router);
   app.use("/api/photos", photos.router);
   app.use("/api/gallery-photos", galleryPhotos.router);
@@ -48,7 +46,6 @@ registerPostProcessors();
 const init = async () => {
   logger.debug("Initialize app start");
   await tokens.init();
-  await stats.init();
   await galleries.init();
   await photos.init();
   await galleryPhotos.init();
